update server to remove string

+++ Server/sklearnhandlers.py
@@ -146,6 +146,6 @@ class PredictOneFromDatasetId(BaseHandler):
             else:
                 newModel(self, DSID)
 
-        predLabel = self.clf[DSID].predict(sample_image_np)[0];
+        predLabel = self.clf[DSID].predict(sample_image_np).astype(int)[0];
         print(predLabel)
         self.write_json({"prediction":predLabel})
